jokes_api/requirements.txt
fastapi
requests
sqlalchemy
mysql-connector-python
python-dotenv
uvicorn
